upgrade to PHPUnit 10
