+++ .gitignore
@@ -15,7 +15,7 @@ error.log
 /tests/test_app/config/Queue/*
 /tests/test_app/config/TestsQueue/schema-dump-test.lock
 /tests/test_app/Plugin/TestBlog/config/Queue/*
-.phpunit.result.cache
+.phpunit.cache
 
 # IDE and editor specific files #
 #################################
