+++ composer.json
@@ -27,7 +27,7 @@
         "cakephp/bake": "3.x-dev as 3.0.0",
         "cakephp/cakephp-codesniffer": "^5.0",
         "enqueue/fs": "^0.10",
-        "phpunit/phpunit": "^9.5.19"
+        "phpunit/phpunit": "^10.1.0"
     },
     "autoload": {
         "psr-4": {
